+++ CHANGELOG.md
@@ -12,12 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Text in GitHub release notes will be copied from here 
   - Some of the version history from GitHub releases has been copied in here
 - QCBOR_VERSION_NUMBER allows > and < version number comparison
+- QCBOREncode_OpenSizedBytes() to encoded definite-length strings in streaming mode
 
 ### Changed
 - QCBOR_VERSION_STRING no longer has "libqcbor" at the start (comply with convention)
 - Removed recursion from negative big number encoding
 - Expand test coverage of big number encoding
 - Source code nits in qcbor_number_encode
+- Improve QCBOREncode_OpenBytes() nesting error detection
 
 ### Fixed
 - Fix encoding of negative big nums that are one non-zero byte followed by zeros.